+++ app/routes/patient.js
@@ -24,29 +24,6 @@ router.post('/:patient_uuid/new/note', patient.note)
 router.post('/:patient_uuid/archive', patient.archive)
 router.post('/:patient_uuid/invite-to-clinic', patient.inviteOneToClinic)
 
-router.all(
-  '/:patient_uuid/programmes/:programme_id{/:view}',
-  patient.readProgramme
-)
-router.get(
-  '/:patient_uuid/programmes/:programme_id{/:view}',
-  patient.showProgramme
-)
-
-router.post(
-  '/:patient_uuid/programmes/:programme_id/clinics',
-  patient.addToSession
-)
-
-router.get(
-  '/:patient_uuid/programmes/:programme_id/new/vaccination',
-  patient.vaccination('new')
-)
-router.get(
-  '/:patient_uuid/programmes/:programme_id/new/ttcv',
-  patient.vaccination('ttcv')
-)
-
 router.get('/:patient_uuid{/:view}', patient.show)
 
 export const patientRoutes = router
